Fix compiler and linker warnings: C4267/C4101/C4834 casts and x64 lib paths

+++ Source/core/dbfile.cpp
@@ -12,5 +12,5 @@ DBFile::Iterator DBFile::begin()
 
 DBFile::Iterator DBFile::end()
 {
-	return Iterator(*this, recordCount);
+	return Iterator(*this, static_cast<unsigned int>(recordCount));
 }
