+++ Source/games/wow/wdc1file.cpp
@@ -77,7 +77,7 @@ bool WDC1File::open()
 	data = getPointer();
 
 	// compute various offset needed to read data in the file 
-	uint32 stringTableOffset = sizeof(header) + sizeof(field_structure) * fieldCount + recordSize * recordCount;
+	uint32 stringTableOffset = static_cast<uint32>(sizeof(header) + sizeof(field_structure) * fieldCount + recordSize * recordCount);
 
 	// embedded strings in fields instead of stringTable
 	if ((m_header.flags & 0x01) != 0)
@@ -89,12 +89,12 @@ bool WDC1File::open()
 	seek(stringTableOffset);
 	stringTable = getPointer();
 
-	uint32 IdBlockOffset = stringTableOffset + stringSize;
+	uint32 IdBlockOffset = static_cast<uint32>(stringTableOffset + stringSize);
 
 	uint32 copyBlockOffset = IdBlockOffset;
 
 	if ((m_header.flags & 0x04) != 0)
-		copyBlockOffset += (recordCount * 4);
+		copyBlockOffset += static_cast<uint32>(recordCount * 4);
 
 	uint32 fieldStorageInfoOffset = copyBlockOffset + m_header.copy_table_size;
 	uint32 palletBlockOffset = fieldStorageInfoOffset + m_header.field_storage_info_size;
